Added combat functionality
Altered Entity class significantly to be more generic.
  - Broke deal damage into a deal and a take function
  - Delegated responsibility for detecting defeat to its own function

The Player class now inherits from the Entity class

The interpreter can now trigger a transition to combat and combat can
trigger a transition back to the inteprter upon its resolution.

Combat can now accuratly detect team one a battle (even if there are
multiple entities on one or more teams)

Made several changes to the way enemies are stored
 - Removed boss library. Bosses are now stored in the EnemyLibrary file
 - Added getEnemy(name) function to EnemyLibrary. It returns an entity
   based on the data for the given name (searching through the enemy list
   and the boss list).

The intepreter now excepts scenes which have only one bit of text as a
string instead of as an array of strings.

Simplified several calls to Text.combatText by adding a default value for
its reset paramter (so you only need to explicitly use a value if you want
to do a reset but not if you don't)